requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
Faker==36.1.1
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
pytest-flask==1.3.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
requests==2.32.3
python-json-logger==3.3.0
marshmallow==3.26.1
Flask-Limiter==2.8.1
flasgger==0.9.5
